# Fix metadata, improve content rebuild performance