--- a/.babelrc
+++ b/.babelrc
@@ -27,6 +27,7 @@
         "loose": true
       }
     ],
+    "@babel/plugin-proposal-optional-chaining",
     "@babel/plugin-syntax-dynamic-import",
     "babel-plugin-macros",
     [