--- a/src/helpers/routes.js
+++ b/src/helpers/routes.js
@@ -1,0 +1,5 @@
+export {
+  buildRoute,
+  buildPathFromFile,
+  normalizeRoute,
+} from "../../buildHelpers/routes";