--- a/src/templates/ApiReference.js
+++ b/src/templates/ApiReference.js
@@ -310,6 +310,7 @@ const ApiReference = React.memo(function ApiReference({ data, pageContext }) {
       <MDXProvider components={componentMap}>
         <LayoutBase
           previewImage={DevelopersPreview}
+          title="Stellar API Reference"
           description="The complete API reference for the Stellar network. Includes descriptions of Horizon endpoints, network concepts, and example code for some languages."
           pageContext={pageContext}
         >