--- a/src/utils/index.js
+++ b/src/utils/index.js
@@ -1,5 +1,3 @@
-import { buildPathFromFile } from "../../buildHelpers/routes";
-
 /**
  * @desc calculates the column width in percentage (%)
  * @param {number} col - { # of columns we want / a total # of columns }
@@ -136,5 +134,3 @@ export const extractStringChildren = (codeSnippet, finalString = "") => {
   }
   return finalString;
 };
-
-export { buildPathFromFile };