I realized that rebuilds when mdx content changes was running extremely slowly, which was frustrating while trying to iterate.

also:
* Fix imports in Jest
* Add optional chaining syntax to babel (`?.`)
* Add some tests to normalizeMdx
* Change how route helpers are imported from site code

Rebuild timings:
before:
```
info changed file at /Users/carlvitullo/workspace/new-docs/content/docs/index.mdx
success createPages - 0.179s
success write out requires - 0.010s
success Re-building development bundle - 0.460s
success run queries - 14.622s - 119/119 8.14/s
```

after: 
```
info changed file at /Users/carlvitullo/workspace/new-docs/content/docs/index.mdx
success createPages - 0.130s
success write out requires - 0.023s
success Re-building development bundle - 0.778s
success run queries - 6.165s - 119/119 19.30/s
```
A little improved, anyway. This might also help with initial build as well.